AURORA: Add a new NCSStack class to replace std::stack in NCSScript
The stack used in scripts doesn't work the same way as a normal stack. We constantly are changing the pointer to the top and the 'base' pointer. The new NCSStack class should get the job done (and without the call to both top() and pop()... (Dr. Seuss?)). Also, implementing the following opcodes:

- o_movsp
- o_savebp
- o_restorebp