MAIN: Fix a stupid crash on Mac OS X
For some stupid reason, eos on Mac OS X crashes when the init()
function is not in the same file with the main.
I'm not entirely sure why, but it may be connected with how SDL on
Mac OS X is set up: It replaces the main...

As a workaround, we'll just move everthing from init.cpp into eos.cpp.
Of course, I'm not entirely happy with that, it adds unecessary
clutter to eos.cpp...