EVENTS: Generically force calling functions in the main thread
Instead of having one request type each for every thing that needs to
be done in the main thread, we now throw functors around with the
parameters bound to what was originally specified.

So we can now have functions that notice they're not being called
from the main thread automagically call themselves in the main thread
with the correct parameters.