NWN: Fix an uninitialized pointer
Even if it was initialized afterward (in redoLayout()), it causes
a severe crash on an OSX system by losing the tooltip adress.

By the way, it seems that in the structure Line in tooltip.h the UString line is
almost unused and redondant with the Aurora::Graphics::Text which already contains
the string.