ALL: Fix the camera move
The camera should not move when the console is open. Since text input
and pressed key are now distinct events, we receive two events when a
key is actually pressed. So we have to test if text input is active.

One feature brings by SDL2 is the ability to recognize a key by is
position and not is name, that is to say we can ask the key next to
tab whatever the layout is (like qwerty, azerty or dvorak).