GRAPHICS: Fix transparency blending
Depth writing needs to be disabled for transparent materials.

The material now has a transparency hint member, ideally
read from the model. If the model does not provide a hint
the transparency is inferred from the textures and diffuse
color. Since Aurora textures compressed with DXT1 are always
opaque, we override the default Ogre claim that DXT1 textures
have an alpha channel.