GRAPHICS: Add back video playing
This time, the video player can be easily adopted to play
several videos simultaneously. This will be necessary
for Sonic.

Right now, we still de-YUV the videos onto a surface held
in CPU memory, and then copy (with pixel format conversion)
it into video memory. This is not optimal and we should
probably change it to de-YUV directly onto the video
memory.