GRAPHICS: Add a new frame locking mechanism
The problem with our current LOCK_FRAME() method that locks Ogre's
SceneMutex is that it blocks in the render loop, therefore producing
a deadlock when locking the frame and then forcing a function into
the main thread (since those functions are handled in the same thread
that's then waiting on the SceneMutex).

To prevent that, we now implement a simple frame lock using atomics
that doesn't block the main render loop. Instead, if the render
method can't aquire the frame lock, it will simply jump over the
rendering and continue with processing the event queue.

To prevent that, we now implement a simple frame lock using atomics
that doesn't block the main render loop. Instead, if the render
method can't aquire the frame lock, it will simply jump over the
rendering and continue with processing the event queue.