GRAPHICS: Add TODO to investigate non-use of parameter
I don't remember what my plans were there. Possible that we need
to re-assign the name to the texture? Or keep it around for checks?
Can that name be different from the texture's resource name?

Is that maybe related to the warning message some generated textures
(from inked PLTs) spit out?