NWN2: Add a Situated class
Mostly copied from NWN.

One difference: NWN2's environmental objects specifiy the position
in a Position struct and the orientation in a quaternion struct.
The former is easily copied, but the latter needs some calculation
to convert them into the roll/pitch/yaw system we use for whole
model rotation.