GRAPHICS: Implement tinting of NWN2 models
Right now, this is done completely in software:
- Go over all model nodes
  - Load the tint texture
  - Create a new, empty texture
  - Decompress the tint texture
  - Iterating over all pixels
    - Read pixel colors
    - Mix tint colors according to pixel values
    - Write the color values into the new texture
  - Stuff the new texture into the TextureManager
  - Append the new texture to the node texture layers

This has, of course, at least two drawbacks:
- It's slow
- Since we're creating uncompressed RBGA textures, it
  takes up a lot of memory, both on the GPU and in system RAM

As a ballpark, the Mulsantir Gates area in Mask of the Betrayer
now takes up 1GB of extra memory, on GPU and RAM each.

Ideally, in the future, this whole tinting business would be
done with shaders instead of manually.