GRAPHICS: Add mesh container class
This adds a mesh container, with the idea of being able to
be shared amongst models (resource sharing). I can be
directly called for rendering, and handles OpenGL buffer
management.

This includes vertex and index buffes updates, and a
wireframe box mesh is included to serve as an example for
default meshes.