GRAPHICS: Texture base class _textureID
Interfaces necessary for OpenGL promoted to the texture base class.
The reasoning behind this is to allow shader samplers to reference
a texture directly instead of only relying on the OpenGL handle.
This easily allows the texture to be reloaded without needing to
synchronise any ShaderMaterial that might be using the texture.