GRAPHICS: Restructure/Rewrite Aurora textures
Specifically, remove the split between PLTs and other textures.
Now, PLTs are full textures, inheriting from the Texture class,
and are handled as such. This removes a lot of unnecessary
complexity from the TextureManager.

Of course, PLT are still dynamic, and need to be tinted
independantly from the engine code. To make that possible
class Graphics::Aurora::Texture now has a virtual method
isDynamic(), which prompts the TextureManager to make the
texture name unique, and not conflate it with textures
using the same base resource.

In addition, TextureManager got two methods,
startRecordNewTextures() and stopRecordNewTextures(),
that lets the engine code make the TextureManager save and
request the names of all newly added texture inbetween
those call. This is essentially a replacement for the old
getNewPLTs() method.