GRAPHICS: A Model can now draw its "skeleton", the node connections
The setting of the line width and point size to these specific values
might not be supported on all platforms though.

Also, the drawing of only the skeleton for the visible model nodes
does not correctly work for the "player" type creature models, the
ones that attach each body part separately. The reason is that we
add these part models as children to the invisible nodes of the
base skeleton.

To visualize this point, this is how it, conceptually, looks:
I---I---I---I---I
    |   |   |
    V   V   V

"I" is an invisible model node, "V" a visible one (only that in
reality, the I and V nodes occupy the same physical place). So
when we're only drawing the visible model nodes, they don't have
any other visible nodes they connect to, ergo no lines between
them.

Does it matter? Dunno. It's just that the whole skeleton for NWN
character is pretty crowded, since there's all kinds of nodes in
there, like the anchor and animation points for the cloak, tail,
wings, weapons, ...