GRAPHICS: Change the way the NCGR grid is calculated
Don't let the maximum width of a column set the width of all cells
in that column, but let the cells "bunch up" instead, on a per-row
basis.

The reason is that main_pnl_off/main_pnl_on are each made up of 13
images, arranged in this fashion:

|     64x32     |     64x32     |     64x32     | 32x32 | 16x32 |
| 32x16 | 32x16 | 32x16 | 32x16 | 32x16 | 32x16 | 32x16 | 16x16 |

Which now can be represented by two rows of each 8 columns, with
the last three of the first row being empty.

Hopefully, we won't need something like that for the rows as well.
And hopefully, we won't see empty cells in the middle of the grid;
those would need an explicit size now.