BUILD: Add configure option --with-release=VER
Used to set the version number suffix, which by default is the git
revision (or "unk", if that can't be evaluated). If called without the
VER argument, the version number is completely unsuffixed.

Note that the REVDESC string, which is part of the full version
number, will not be modified by this.