KOTOR: Add code for testing KotOR animations
Highlight a creature with the mouse, and press m to advance this
creature to the next animation. A message showing the name of the
current aniamtion will be printed to stderr. Press n to go back to the
previous animation.

Likewise, highlight a situated object (a placeable or a door) and
press to advance to the next object state.

This only works with objects and creatures that are already
highlightable.