AURORA: Change how GFF3Struct::getLocString() works
It now doesn't throw on failure, but returns false. It also keeps the
original contents of the LocString its passed in that case.

This makes getLocString() easier to work with, especially with how
we're using it in the engines. We don't need to wrap it in a try-catch
block anymore and precede it with a hasField().