GRAPHICS: Render environment maps in ModelNodes
Works for KotOR and KotOR2.

This now requires models with environment maps to be rendered three
times, to have the environment map blend correctly into the model.
Once the OpenGL graphics code is rewritten to use the shader pipeline,
this will be done in one pass in a shader instead.