GRAPHICS: Add a TextureMode parameter to TextureManager::set()
The TextureMode parameter lets the calling code specify what to use a
texture for. Currently possible values are kModeDiffuse, for the usual
diffuse texture, and kModeEnvironmentMapReflective, for a reflective
environment map.

This effectively decouples the isCubeMap() texture property from its
usage, letting you use a non-cubemap as a simple reflective
environment map. That is used, for example, in the Canian wasteland in
the Hordes of the Underdark expansion of Neverwinter Nights: the icy
glacier faces use a simple two-dimensional texture for its
reflections.

Unfortunately, those glaciers don't yet renderer correctly in xoreos;
they're far too bright. It's possible that the render method is
different from what KotOR does.