AURORA: Add environment variables to NCSFile
Similar in effect to how the environment works in UNIX shells, the
environment is accessible throughout the run of a script. A child
script will inherit the environment of its parent script.

The immediate use case is event types in Dragon Age scripts: objects
in the Dragon Ages games have each only one script, that is called
with an event type variable specifying what it's been called for.
Within the script, the GetCurrentEvent() function retrieves the event.

Likewise, the userdefined scripts in other games use an event type
work very similarily.