AURORA: Implement NCSFile::o_writearray
Offset to the array on the stack is a direct argument. The index is
the top of the stack, followed by the value to write. The index is
popped off, and the array value is updated, growing the array if
necessary. The value written is left at the top of the stack.