DOCS: Re-sort the Contributors list
Sorting by last name first doesn't make any sense when we don't have
real names for everyone (and there's no reason why we should require a
real name), and this notion of a family name that's a last name isn't
even true for all real names.

String sorting names in general is a bit awkward, because sorting
order depends on the locale.

So we're just stating them by order of first commit now.