GRAPHICS: Fix Model::getTooltipAnchor()
Since we removed the automatic world re-rotation, the Z axis is again
the going from the ground into the sky. This means getDepth() returns
the height of an object model, and we need to take that into account
when calculating the tooltip anchor point.

Also, remove the fudge factor. We need to do that in pixel space.