COMMON: Give each debug channel its own verbosity level
This completely rewrites how the DebugManager handles debug channels.
Each debug channel now has an associated verbosity level, independent
of the verbosity level of other channels.

Additionally, the engine debug channels are now created by the
DebugManager as well, instead of having to be registered at engine
start. This removes a bit of the awkwardness related to --listdebug
not being able to print them. On the flip-side, they're now a bit less
flexible.