SHADER: Flag based shader builder
ShaderBuilder takes a set of flags and constructs an appropriate
shader source string. A name can also be generated, and is used
by ShaderManager to determine if flags indicate a shader that has
already been built.

Because of threaded loading, all shaders are still built at
initialisation time. The appropriate shader object is simply
selected by using the construction flags. If a particular
combination of flags has not been initialised, it's likely that
either an error will occur, or a shader program id of zero is
used (depending on video card drivers).