SOUND: Capture channel count and sample rate in the MP3 decoder
Instead of having getChannels() and getRate() always read the
mad_frame structure, query this information once in the constructor
and save it in member variables.

This gets rid of a crash in a rare edge case: when the number of
samples buffered in the sound update thread fits evenly with the
number of samples in the file (either directly or through looping),
the mad_frame structure will have been wiped after reading these
samples. Further calls to getChannels() and getRate() will then return
0 until the stream has been rewound.

This edge case happened about 4 minutes into playing/looping the
environmental sound bed_852nih in the module 852nih of KotOR2.

Thanks to Heavylobster for finding this edge case.