GRAPHICS: Use >= to check for end-of-frame when interpolating
Just to make sure we're not accidentally overshooting the keyframe.

Maybe we should also take into account the next keyframe when
interpolating, so that the next position/orientation can be used
instead of "stopping" at the last frame or continuing with the current
direction?

Also, currently, some animations interpolate weirdly. For example, the
tlkplead ("Talking, pleading") animation has the hands spin wrongly.
This does not show up when disabling interpolating. This could be
related to interpolating quaternions for the orientation keys. It
seems that a simple linear interpolation, like we do currently, is not
correct. SLERP, spherical linear interpolation, seems to be the way to
go instead.