GRAPHICS: Fix quaternion interpolation for certain angles
When the angle is > 90°, we need to flip the direction of the second
quaternion to get a smooth transition instead of wild jumps.

This fixes the hand spinning on the "Talking, pleading" (tlkplead)
animation, the arms wobbling on the worship animation and the
flickering when opening a chest or armoire. And possible other issues,
too.