GRAPHICS: Overwrite duplicate nodes in NWN model
When we encounter a duplicate node in an NWN model, delete that node
(and fix up the child/parent relations) and let the new node instance
take its place.

This is necessary for at least the "c_asabmage" (Asabi) model, which
has two instances of the "Rbicep_g" node. Without this overwriting,
the animations for the right arm aren't displayed correctly, with the
upper arm not being connected to the lower arm. This is visible in the
cutscene for the Chapter 2 of the first NWN expansion, SoU.