GRAPHICS: Update the model bounding box on state/animation changes
This makes the bounding boxes more accurate, especially for scaled
creatures like halflings and dwarves.

We're only updating the box for the first frame of the animation,
though. Doing it continuously, without caching at least, is relatively
slow. In the entry area of Chapter 2 of the NWN OC, for example, this
would create a decrease of 20 FPS for me.

Since position and orientation animation keys are separate and per
node, caching the bounding box for every model animation frame is also
non-trivial.