NWN: Check dynamic_cast<Foo *> for 0 before dereferencing
dynamic_cast<> can fail! This happens when the source object can't be
transformed into an object of the target type (because, for example,
they don't share one inheritance tree).

In the case of dynamic_cast<Foo *>, it may return 0, so we need to
check for 0 before dereferencing there. Otherwise, we might
dereference a 0 pointer.

In the case of dynamic_cast<Foo &>, it may throw an exception. If we
have a pointer that we know isn't 0 and are sure the cast *should*
work, we can use this version of dynamic_cast<>, too.

This fixes Coverity Scan issues #1354909-#1354913
and #1361965-1361969.