BUILD: Link the engines into the main instead of into libengines.la
Now that the cyclic dependency is gone, the engines do not need to be
needlessly copied into libengines.la anymore. Instead, the can be
linked in directly with the main xoreos binary.

This cuts down on 400MB of (unstripped) object copies, which would
have been just written and immediately read again.