MAIN: Always properly clean up after ourselves, even on error
We may still add a really fatal error category that shouldn't try to
clean up, but all the usual errors we currently raise are not that.

This also gets rid of spurious OpenAL warnings and potential segfaults
related to OpenAL memory being destroyed automatically by the OS while
a separate thread still initializes it. If we're calling the OpenAL
cleanup functions ourselves on quit, they should do the right thing
and wait for locks, etc.