GRAPHICS: Rework material construction for ModelNodes
Building materials for use wtih ModelNodes is reworked to only be
performed if a dirty flag is set, and then only when the modelnode
is actually attempted to be rendered.

All data must be ready by render time, so this ensures that only
those nodes shown will have materials, and it can be certain that
everything else is loaded and set.