NWN: Correct the scrolling math for WidgetGridBox
It now displays the correct number of items in the last row.
Previously, instead of displaying the last N items, it tried to
display the last ItemsInARow - N items. This lead to some items
potentially not being displayed at all

Also, it now doesn't crash when scrolling by dragging the scrollbar
and the items fill out of the last row completely. In this case, the
old code calculated the starting item wrong, being off by the number
of items in the row (i.e. one row).