GRAPHICS: Fix GL_UNPACK_ALIGNMENT for 24-bit and 16-bit textures
Even for power-of-two textures, we can't really tell the driver that
the texture rows are 4-byte aligned for textures with 3 or 2 bytes per
pixel. Because they might very likely not be.

We *could* do it for some mipmaps, but we only set GL_UNPACK_ALIGNMENT
per texture, not per mipmap. Would it be worth it doing it per mipmap?
No idea.