GRAPHICS: Render fonts using shaders
Render fonts (texturefont only right now) using shaders. It's not
the most efficient method, but right now just removing older GL1.x
code is more important.

Gui rendering is done in "immediate mode", i.e not put into a render
queue. This is for reasons of the render queue not (yet) being able
to handle something that doesn't fit within the shaderrenderable
paradigm.

Fonts are always white. This needs to be fixed. And a whole bunch
of compiler warnings removed. And some of the font rendering code
needs to be cleaned up. Think of this commit as a backup snapshot.