COMMON: Use boost::atomic to signal thread running and killing
Otherwise, this might be prone to race conditions. volatile isn't the
correct way to deal with this, either.

We can't really force a thread to die, either. SDL1 had a function for
that, but it was not portable and unreliable, so they removed it.

We can, however, detect that the thread is still running and issue a
warning.