GRAPHICS: Fix Xbox KotOR/KotOR2 model animations
The Xbox versions of the KotOR/KotOR2 models store animations a bit
differently than the Windows/Mac/Linux versions:
1) 8 vs 12 reserved bytes before the mapping
2) Bone indices are stores as int16 instead of float

This commit makes our model loader read them correctly. To correctly
identify whether it's an Xbox model or not, a parameter Xbox is added
to the constructor (similar to how we differentiate between KotOR1 and
KotOR2), set by the engine.