JADE: Add support for installations with multiple languages
There are several versions of Jade Empire with multiple languages:
- The GOG version before the October 2017 update that split the
  language builds
- The Xbox versions
- Aspyr's Android port

All of these come with sound files in only one language, but multiple
languages for text.

xoreos now correctly identifies all available languages in a Jade
Empire installation and lets the user select one. This is *not*
changeable while the game is already running, though.