COMMON: Fix INFLATE without size when size evenly divides frame size
When decompressing without knowing the output size beforehand, we
decompress into buffers of frameSize length and then paste them
together into one output buffer at the end.

When pasting, we used to calculate the size of the last frame with
"% frameSize", which, if the output size evenly divides the frame
size, is of course 0.

We could just check for 0 there and use the frame size instead then,
but using a running counter of bytes left to write is easier.