TESTS: Add unit test for Engines::Trigger class
This test will pass an array of points to a utility Trigger
class that define a close, non-convex polygon. The class'
'contains(float x, float y)' function is called for a set
of test points. If the function is working properly, it will
return the expected boolean result declared in the test.

See the tests/engines/trigger.cpp comments for illustrations.