NWN2: Skeleton framework for Trap class
Create an initial Trap class definition by including NWN2 script
function calls as well as variables set in the .gff file for traps.
This class will be inherited by Door, Placeable, and Trigger classes;
it is not intended as a stand-alone object. The class includes calls
to process detect, disarm, and trigger events on the trap. A Trap
class instance can be created either from a GFF3Struct or by a
Creature class instance using a data row in the traps.2da file.
The latter is intended for a trigger region created from a trap item
or spell.

Note: This version of the trap.cpp file compiles with multiple
warnings for unused variables.