NWN2: Implement Factions class
Factions are stored at the module level of the game, so the
Factions class is added to the Module class and accessed by the
Object class via the _module variable in the Area class. A faction
ID is assigned to a Creature, Door, Placeable, and Trigger object,
with the faction vs. faction reputations determining whether a
Creature or Trap behaves in a friendly, neutral, or hostile manner.
(Traps will not trigger vs. a friendly faction, for example.)

Creatures can store additional reputation information for
individuals that can be modified via script calls. This is held by
a Reputation class. Each creature stores how other creatures feel
about them, with the default being the faction reputation.

Initially, Faction data is just retrieved from the repute.2da data
file. The door and placeable classes will check the faction
reputation before determining whether to trigger their trap.