APPVEYOR: Revert to previous image of AppVeyor
There's two issues with the current image:

1) vcpkg is outdated and fails to install anything. Also happens in
   other projects: https://github.com/OSGeo/gdal/issues/1747 . We can
   fork the vcpkg repo and compile it ourselves, then vcpk itself
   works.
2) The current OpenAL-Soft recipe is broken. So even if we fix 1),
   it's stil broken. The relevant vcpkg issue is tracked here:
   https://github.com/microsoft/vcpkg/issues/7790

So for now, we're downgrading the AppVeyor image to the previous one,
and this seems to fix the build. For now, at least, until the next
AppVeyor update and then the current one is the previous one.
Hopefully, there's a proper fix until then.