GRAPHICS: Tell DearImGui to not mess with the mouse cursor
We're already handling it ourselves. Specifically, we want to not hide
the mouse cursor if no alternative cursor has been defined, which
happens for all the console versions.

In the future, when we're actually using ImGui for stuff, we might
need to see how that'll work out when interacting with ImGui widgets.