GRAPHICS: Allow lights to be registered and activated
Lights can be registered and deregistered to the light manager.
Registered lights within a position and radius can be listed and
loaded into shaders for rendering. NWN used as the test game.
Light positions are not always correct.

Calculate modelview matrix internally within camera manager to
make the modelview matrix readily available to the light manager.

Signed-off-by: mirv <mirv.sillyfish@gmail.com>